Bug 551740 – unable to manually adjust the clock past 13:49:49
GtkAdjustment behaviour has changed to restrict the range of the values.
From the GTK README:
  GtkAdjustment now enforces that values are restricted to the
  range [lower, upper - page_size]. This has always been the
  documented behaviour, and the recommended practice is to set
  page_size to 0 when using adjustments for simple scalar values,
  like in a slider or spin button.
This change meant that the time couldn't be set to greater than 13:49:49, as the upper was 23:59:59 and the page_sizes 10:10:10.
This patch reduces the page_sizes to 0, as recommended in the above quote.
Thanks to Sebastien Bacher for finding the above information.