Bug 580182 – Don't allow creating groups overwriting existing ones
When creating an user, don't allow its main group to use an already existing name, which overwrites the group and its members. This was a security issue when destroying admin groups. Thanks to Jonathan Davies for the original patch.

A trivial fix also restores the check that prevents a new group to be named as an already existing one using the "Create group" dialog.