Port to PolicyKit1
This removes most usage of that lib on the client side. The backends now take care of most details, and GstPolKitButton is replaced by PolkitLockButton from libpolkit-gtk, which is now the only lib we link to. Add a requirement on the system-tools-backends >= 2.8 to be sure users don't mix Policykit versions.