Fix broken time zone combo box (GtkBuilder 2.16 update)
In the move to GtkBuilder files, we lost the models, which was leading to an empty list and a crash. Now we create the model from the .ui file again, but still create the cell renderers from code. The .ui file changes a lot due to Glade cleaning things. See https://bugs.launchpad.net/ubuntu/+source/gnome-system-tools/+bug/406211 for the crash.