When checking for main group's existence, take into account the combo box selection
We now check that the main group to create does not already exist, to avoid overwriting it. This commit uses the group selected in the user_settings_group combo box instead of always using the login for that. This allows choosing another group, but has little practical interest until we actually allow using existing groups without replacing them.