Allow new users to explicitly choose an existing main group
The previous checks were so restrictive that you could not choose an existing group as main group for new users. Two issues have led to that: the old behavior was sheerly overwriting groups, raising to security problems; users were not added to their main group if it existed. Now, we consider that automatically adding an user to an existing group because it's named as its login is dangerous, so we avoid that. But we accept using any existing group if it has been manually chosen in the list. We also add users to their main group if needed. After the freeze, error messages should be improved and a dialog could ask before using system groups as main groups for a new user.

Also add a new group_settings_get_group_from_name() function, which is almost identical to what group_settings_group_exists() did before, and use the former from the latter.