Merge all main group management in set_main_group()
Before we were getting it in get_main_group(), and setting them after; this required two separate commits (one to create the group, one to add the user to it). This removes one error in case it fails, and better error handling is performed by returning NULL, allowing to cancel the user creation. The old set_main_group() function is renamed to the more accurate select_main_group().