Don't commit group changes when not required (self user changes)
It's useless since modifying the self user should not lead to changing any group. This required to get .set rights, which are reserved to administrators, thus raising an error after succeeding to modify self. Also always check that we actually need to commit changes (new group, add user to group) before doing so.