Force listing system groups
The GConf "showall" option applies to both system users and system groups. This was not working in previous releases because of some hidden bug, and we actually need to show system groups independently of system users. This is a temporary hack to fix a regression, waiting for a more complete change in master.