Fix handling of services
Use a new feature in liboobs that allows passing priority 0 to either keep the previous value, or use the default. This fixes scripts being reset to priority 50 when re-enabled. Also set services as either started or stopped for a given runlevel, instead of using OOBS_SERVICE_IGNORE. This is the correct way of disabling scripts, and allows us to store the priority in the file name.