Rework user profiles management
Rework GstUserProfile(s) so that we can detect what profile as given user corresponds to, according to its settings and group memberships, which are considered *as opposed to other profiles*. This allows for manual tweaking of group memberships without losing the profile identification. Most privileged profiles have to go to the end of the configuration file, so that users that match several of them get the most powerful. The profiles member of GstUserProfiles goes from GHashTable to GList because of that new ordering significance.

This leads to create gst_user_profiles_get_for_user(), gst_user_profiles_apply(), and gst_user_profiles_get_from_name(). user_settings_find_new_uid() is no longer static so that we can call from user-profile.c; it should eventually move to liboobs. Obsolete code setting default profile for user_settings_dialog is also removed.