Support user profiles descriptions
Profiles configuration file can now provide a description that is displayed in a label associated with radio buttons. For this gst_user_profiles_get_names() becomes _get_list(), and simply returns a pointer to the internal profiles list, so that we can access names and descriptions. Add Pango to the dependencies as we now access it directly.