Improve checking of digits from Real Name
Avoid adding login proposals in the list that start with a digit, as they would be rejected when chosen. Don't check that the first char of the real name is not a digit: instead, allow for this, but only add login proposals that would be correct (e.g. when starting with the second word).