Don't commit changes when all user deletions have been canceled
We support deleting several selected users at the same time. But we need to check how many of them will finally be deleted, in case the user has chosen cancel for all of them. This is especially an issue now that PolicyKit authentication dialog may be shown when committing: if only one user was selected, and its deletion has been canceled, asking for authentication seems to indicate that the user will be removed anyway.