Bring back checks for password and advanced user settings
Use the checks again for the advanced and the password dialogs. Checks are performed when closing the dialogs for now. Remove the now obsolete user_settings_dialog_run(), and split some tests from the test battery when needed.

For password, remove the "changed" property of a simpler mechanism: when the password is empty, no changes are performed. This behaves mostly the same from the user's point of view. Improve home dir checking, warning the user if it already exists that permissions won't be changed.