Prevent deleting accounts of active users
We were nicely asking the user to choose what to do about active users, but deluser always fails with logged in users. Thus, only show a message dialog explaining the problem instead. Anyway, it's wiser not to remove accounts while people are working on them, since it could lead to strange results.