Add checks when removing admins or admin rights
When deleting an user, check that it's not the only amdin account left. If that's the case, show an error dialog and block deletion.

When removing admin rights from an account, warn if the selected user is the current one (i.e. the one corresponding to OobsSelfConfig). Refuse to remove the admin rights of the last admin left on the system. This is supported either when moving from a profile implying admin membership to a non-admin one, or when unselecting the admin privilege in the privileges table. People can still lose admin rights when tweaking groups membership, but they must know what they are doing...