Move to new users and groups STB protocol
The new protocol in the System Tools Backends >= 2.9 (liboobs 2.29.2) no longer commits users config as a whole. Instead, we use oobs_user_commit() and oobs_users_config_add/delete_user(). Groups are still committed as a whole when changing a user's privileges, but individually in the groups dialog.

Keeping references to OobsListIters for groups in groups and privileges table is now useless, making the functions simpler.