Add user profiles configuration file
This file is used by GstUserProfiles, but only shipped by distributors. Include it in the tree, waiting for packagers to upstream the changes they need (groups). This version is veyr limited: we only define groups for two profiles, other settings will be chosen by the platforms.

Move this file to $(sysconfdir)/gnome-system-tools/user-profiles.conf.