Update password status field
Fill the "Passwd:" field depending on whether the password is empty and asked on login, which is not yet filled by the backends. Call user_settings_show() after committing, since password may have changed. Reduce padding and modify table so that the label can take more horizontal space without changing the size of the users list.