Preliminary support for disabled accounts
If you create an account and close the password dialog without choosing one, new account will be disabled. To avoid confusion, show a warning with a button to enable it. Also add a check box in the Advanced dialog to Enable/Disabled accounts. All these options are hidden for now, waiting for actual support from the backends.