Improve services list contents and appearance
Now that the backends filter dangerous/useless services, we can show all they send us, even if they are not referenced. In this case, only show the service name and generic icon.

Use the short name of the service in bold, and the short description below; long description, when present, goes in a tooltip. This avoids going beyond the dialog's length is is much easier to read. Use 32px icons instead of 48px to save much vertical space, without losing any information. Reduce padding around checkboxes to 6px.