Allow gst_tool_add_configuration_object() to skip updating child objects
Since OobsUser and OobsGroup are now individually committed, we add them as configuration objects. But we don't want them to be updated individually: this triggers too many re-reads of the configuration in the backends (admittedly a bug), while simply updating OobsUsersConfig and OobsGroupsConfig handles destroying old child objects and creating new ones. So add an additional parameters, only used in those cases, to only watch "commit" signal, and not "updated".