Enable account password (un)locking
Support for this is in the backends 2.9.4. This is particularly useful for new accounts for which no password has been set, in which case accounts are disabled, so we need a visual notice and a way to enable them. Enabling accounts with an empty password requires setting a new password, which is not supported through the Advanced Settings dialog ATM.