Fix hang when changing empty password to something else.
If user has an empty password, authentication phase is skipped. In that case, calling io_queue_pop() will print current password when 'passwd' is waiting for the new one. This is due to the asynchrous behavior, which means that io_watch_stdout() is only called after this mistake is done. So we just wait for io_watch_stdout() to be run and print the new password from there. This is just one more hack into this very tricky code.