Migrate to GSettings
Remove gst-conf.{ch} which is now useless, since it was only wrapping getters and setters around GConf's API. Create a GSettings object in GstUsersTool, since it's the only tool to use settings. Migrate schemas and update configure.in and Makefile.am accordingly.

Raise dependency on GLib to 2.25.3 (needed for GLIB_GSETTINGS m4 macro), and make GIO a "core" dependency (it was separated most likely for historical reasons).

intltool support is not yet here, so we'll have to wait for keys' descriptions to be translated.