Don't limit users list's width request
The VBox width request was set to 200 to keep it from being either to wide or not enough (to enable ellipsizing). But this was preventing the buttons from growing when their labels needed to: result was that at high font size and with some languages, the text was getting out of the buttons.

Setting the width request for the ScrolledWindow of the users list fixes the problem, letting the button grow as they want.