Make groups management UI clearer
It's been reported that the Add/Delete buttons of the
groups management dialog are ambiguous, because you
don't know whether they add a group to the system or
the user to a group. This was a problem esp. in Chinese.

Move the Manage Groups button so that it's obvious that
you're not editing settings for the current user, but
for the system. Also explain that the groups list shows
available groups, and not groups the user is member of.

https://bugs.launchpad.net/ubuntu-translations/+bug/603746