Improve support for changing home directory
When changing the home directory path, ask whether to use
the new dir's contents, or to migrate data from the old one.
In each case, the old dir can be deleted or kept, and the dir
can be chowned to the user.

This requires support for home dir flags in the backends,
that are only present with 2.10.1 (else it has no effect),
and in liboobs 2.31.91.

We also need to commit changes from the advanced settings dialog
async, because copying files can be slow, and we'd reach D-Bus timeout.
Changes in liboobs and the backends introduce a very long timeout for this.