Fix interaction of password-less login with disabled accounts
A disabled account shouldn't be allowed to be in the password-less
login group, since that creates unexpected effects of user being
allowed to login via GDM, but not from the console.

The most problematic case is when creating the user, keeping an
empty password, and checking the password-less login checkbox:
we report the account is no longer disabled, which is wrong, and
create the weird situation described above.

So when disabling accounts, remove them from the password-less
login group. When changing password, don't unlock account if no
new, non-empty password was provided, and refuse to enable password-
less login. In the future, an error message or some GUI fix would be
better.

See https://bugs.launchpad.net/ubuntu/+source/gnome-system-tools/+bug/630430