Improve performance when filling tables in users-admin
It's absurd to get the model from the tree view again and again
when inserting users or groups: doing this for about 150 objects
has a real performance hit on startup. So store models as a static
variable in each file (OOP of the poor). Making it a GtkListStore
allows saving a few casts to GtkTreeModel in the insert functions
that are intensively used.

Also stop adding users that aren't shown to the table: they are most likely
useless. If settings change and we are asked to show system users, we can
reload the whole UI anyway. Same for privileges table, where a
GtkTreeModelFilter was absolutely useless - remove them.

Finally, use gtk_list_store_insert_with_values() to avoid running the
filter/sort functions on empty rows.