Update groups after creating or deleting users
Previous behavior was clearly broken, since groups were not realoaded,
meaning that new user's main group wasn't known. When committing further
changes to new users, an invalid GID was passed, and every command failed.
We can also stop reloading groups when running the groups dialog, which was
a mere workaround.

This requires a change in liboobs to work correctly, so bump dependency to
2.91.1.