Fix libemap.a linking with --add-needed
When linking with the --add-needed option, ld doesn't find symbols
from -latk-1.0 because they are specified before libemap.a, while
libs containing functions used by an object are expected to appear
after it. This leads to this error when trying to resolve
atk_object_set_name(), which is used by libemap.a:
/usr/bin/ld: /usr/lib/libgtk-x11-2.0.so: undefined reference to symbol 'atk_object_set_name'
/usr/bin/ld: note: 'atk_object_set_name' is defined in DSO /usr/lib/libatk-1.0.so so try adding it to the linker command line
/usr/lib/libatk-1.0.so: could not read symbols: Invalid operation

This was uncovered by Ubuntu when switching their builds to --add-needed.

Change time_admin_DEPENDENCIES order for consistency.